# api/openapi.yaml
openapi: 3.0.3
info:
  title: DevSandbox API
  version: 1.0.0-alpha.1
  description: Public management and connectivity API for DevSandbox.
servers:
  - url: https://api.devsandbox.example
security:
  - bearerAuth: []
...
paths:
  /v1/auth/device/start:
    post:
      tags: [Authentication]
      operationId: startDeviceAuthorization
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStartRequest'
      responses:
        '200':
          description: Device authorization started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStartResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/auth/device/poll:
    post:
      tags: [Authentication]
      operationId: pollDeviceAuthorization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicePollRequest'
      responses:
        '200':
          description: Device authorization status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePollResponse'
        '202':
          description: Authorization remains pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePollResponse'
        default:
          $ref: '#/components/responses/Error'
  /v1/me:
    get:
      tags: [Identity]
      operationId: getCurrentUser
      responses:
        '200':
          description: Authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: DevSandbox session
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - invalid_request
            - authentication_required
            - forbidden
            - not_found
            - conflict
            - invalid_state
            - quota_exceeded
            - provisioning_failed
            - transport_failed
            - internal_error
        message:
          type: string
          description: User-safe diagnostic message.
        details:
          type: object
          description: Optional non-secret structured context.
          additionalProperties: true
    DeviceStartRequest:
      type: object
      additionalProperties: false
      properties:
        clientName:
          type: string
          maxLength: 100
    DeviceStartResponse:
      type: object
      additionalProperties: false
      required: [flowId, userCode, verificationUri, expiresAt, pollIntervalSeconds]
      properties:
        flowId:
          type: string
        userCode:
          type: string
        verificationUri:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        pollIntervalSeconds:
          type: integer
          minimum: 1
    DevicePollRequest:
      type: object
      additionalProperties: false
      required: [flowId]
      properties:
        flowId:
          type: string
          minLength: 1
    DevicePollResponse:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state:
          type: string
          enum: [pending, authorized, denied, expired]
        sessionToken:
          type: string
          writeOnly: true
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      additionalProperties: false
      required: [githubUserId, githubLogin]
      properties:
        githubUserId:
          type: string
        githubLogin:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
...